refactor(processor): extract text utilities
Move normalization, rounding, and text dedupe helpers out of the main processor so the page extraction path can be split more cleanly.
